% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{addto_pipeline}
\alias{addto_pipeline}
\title{Add script specific folders to pipeline}
\usage{
addto_pipeline(name = NULL, ...)
}
\arguments{
\item{name}{Character, name of script without .R extension}

\item{...}{Additional arguments}
}
\value{
Creates folder in project structure and assigns global path to
script sub folder. Returns a message if successful.
}
\description{
\code{addto_pipeline} creates a folder inside the top-level pipeline folder,
ensuring an easy association between .R files and the output from them.
A path variable to the folder is also assigned to the global environment.
}
\details{
A top-level folder named pipeline is created if it does not exist already.

Robust function: TRUE
}
\examples{
\dontrun{
addto_pipeline(name = "my_script")
}
}
