% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pipeline.R
\name{clean_pipeline}
\alias{clean_pipeline}
\title{Delete empty pipeline folders}
\usage{
clean_pipeline(dir_path = NULL)
}
\arguments{
\item{dir_path}{Directory path}
}
\value{
Message on which, if any, folders were deleted.
}
\description{
\code{clean_pipeline()} differs from \code{tidy_pipeline()} in that the
latter \emph{moves} all pipeline folders without a matching .R file to the
\emph{_archive} sub-folder, while \code{clean_pipeline()} deletes all pipeline
folders that are empty. Such folders typically exist if a script does not
result in any output.
}
\details{
\emph{Not very elegant, repeating code, should be rewritten}
}
