% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{create_project_folders}
\alias{create_project_folders}
\title{Create folders for different parts of academic project}
\usage{
create_project_folders(proj_folder = NULL, ...)
}
\arguments{
\item{proj_folder}{Path to .Rproj folder}

\item{...}{Other arguments}
}
\value{
Message if folders are created
}
\description{
\code{create_project_folders()} is a function that within the project folder
creates a set of sub-folders. Currently: data, pipeline, R, manuscript,
output and lit. Within each of these it also creates an _archive folder.
In the manuscript folder a sub-folder called latex is also created to hold
.tex files to include in markdown YAML.
}
