% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_wrangle.R
\name{del_main}
\alias{del_main}
\title{Delete files created by knitr and markdown when something goes wrong}
\usage{
del_main(directory = NULL)
}
\arguments{
\item{directory}{Character vector with path to where the error files are
stored.}
}
\value{
Deletes the files, nothing is returned
}
\description{
\code{del_main} is a function that deletes output files created for
debugging errors when rendering the paper using knitr and rmarkdown.
This will also delete _main.tex and _main.pdf. Function should only be
used after something has gone wrong with compiling the paper.
}
