% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Extract file path from folder}
\usage{
get_file(name, ext = "fst", file_num = 1)
}
\arguments{
\item{name}{Character vector of the name of the file to look for}

\item{ext}{The file extension \emph{without} a period (.)}

\item{file_num}{Numeric. In case more than one file matches the \code{name}
argument, one can specify which of the files to return. Defaults to one.}
}
\value{
Returns the file path, if the file is not found it returns an error.
If more than one file matches the name, a list of all matches is
also printed to console to facilitate picking the right file_num
}
\description{
\code{get_file} is a function that takes the name of a file and returns
the path to the file, if found.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
