% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_paper.R
\name{open_paper}
\alias{open_paper}
\title{Open a pdf}
\usage{
open_paper(filep = NULL, pdf_viewer = NULL)
}
\arguments{
\item{filep}{Character vector with path to file}

\item{pdf_viewer}{Path to .exe file for pdf program to open PDF with}
}
\value{
Opens pdf
}
\description{
\code{open_paper} opens a pdf file to view in pdf reader program
If the output from rendering the paper with RMarkdown is located somewhere
other than project directory, the pdf will not open automatically.
}
