% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_wrangle.R
\name{safe_rmd_prevw}
\alias{safe_rmd_prevw}
\title{Preview chapter from other location than project folder}
\usage{
safe_rmd_prevw(chapter = NULL, folder = NULL)
}
\arguments{
\item{chapter}{Name (with .Rmd extension) of chapter to render}

\item{folder}{Path to where .Rmd files are located within project}
}
\value{
PDF
}
\description{
\code{safe_rmd_prevw} allows the user to preview a book chapter
rendered with \code{bookdown} when the chapter .Rmd file is located in
a sub-folder away from the project file. It creates a temporary environment
and executes the \code{bookdown::preview_chapter()} within this temporary
working directory. Furthermore, it stores the output pdf in the same
directory.
Robust function: FALSE
}
