% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_wrangle.R
\name{save_version}
\alias{save_version}
\title{Save paper draft}
\usage{
save_version(folder = NULL, title = NULL, note = NULL)
}
\arguments{
\item{folder}{Directory to look for manuscript}

\item{title}{Paper title, if NULL puts project name in front}

\item{note}{Character to append to filename}
}
\value{
Returns the new file path invisibly
}
\description{
\code{save_version} saves a copy of the manuscript
with a custom note appended to the filename by the variable \code{note}.
If no note is given the filename is extended by a version number.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
