% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{start_project}
\alias{start_project}
\title{Initiate R project with defaults}
\usage{
start_project(dir = NULL, proj_name = NULL, ...)
}
\arguments{
\item{dir}{Path to where project folder should go}

\item{proj_name}{Character, name of project}

\item{...}{Other arguments}
}
\value{
Message if successful
}
\description{
\code{start_project()} is a function that does three things:
\enumerate{
\item Relies on \code{usethis::create_project()} to setup a new R project.
\item Calls on \code{create_project_folders()} to populate project with
relevant folders.
\item Writes a setup script using \code{write_rsetup()}
}
}
