% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{tidy_pipeline}
\alias{tidy_pipeline}
\title{Tidy pipeline project folder}
\usage{
tidy_pipeline(rpath = NULL, path_pipeline = NULL)
}
\arguments{
\item{rpath}{Path to project folder with R scripts}

\item{path_pipeline}{Path to pipeline}
}
\value{
Message when folders are moved
}
\description{
\code{tidy_pipeline()} moves pipeline folders that have
names without a matching counterpart in the R/ project folder.
}
