% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymess.R
\name{tidymess}
\alias{tidymess}
\title{Long user messages in a tidy way}
\usage{
tidymess(..., type = "warning", prefix = " ", initial = "")
}
\arguments{
\item{...}{Message to user}

\item{type}{Character, either "error", "warning" or "message"}

\item{prefix}{Prefix before message}

\item{initial}{Initial, ""}
}
\value{
The input message either in form of an error, warning or message.
If the input type does not match either, an error is thrown.
}
\description{
\code{tidymess} let's the programmer write longer and more informative
messages to the user by ensuring that the message is printed in a tidy way
to the console. One can specify which type of message to give: error,
warning or message.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
