% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{write_rsetup}
\alias{write_rsetup}
\title{Write setup script to R/ folder in project directory}
\usage{
write_rsetup(proj_folder = NULL, packages = NULL, ...)
}
\arguments{
\item{proj_folder}{Path to .Rproj folder}

\item{packages}{Character vector of packages to include}

\item{...}{Other arguments}
}
\value{
Writes an R file to R folder and gives message if successful
}
\description{
Write setup script to R/ folder in project directory
}
